backend/requirements.txt
# FastAPI + ASGI server
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.1

# Needed by pydantic-settings for .env file loading
python-dotenv==1.0.1

# ERA5 data access (on-demand Zarr computation)
gcsfs==2024.3.1
zarr==2.18.2
xarray==2024.5.0

# Numerical processing + tile rendering
numpy==1.26.4
Pillow==10.3.0
mercantile==1.2.1

# Google Cloud Storage for tile caching
google-cloud-storage==2.17.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
